gpu_server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
Pillow==10.3.0
transformers>=4.40
accelerate
torch>=2.1
diffusers>=0.27
structlog==24.1.0
